requirements.txt
discord.py
asyncio~=3.4.3
pytz
datetime~=4.3
pokepy~=0.6.1
pandas~=1.1.3
pyenchant~=3.2.0
pytablewriter~=0.58.0